requirements.txt
discord.py
python-dotenv
alpaca-py
